Adding the device specific README
[Description]
This commit adds the README for the am62x & am64x device. The README
contains the `List of examples`	with the SDK versions on which it was
developed and tested.

The top level README is also updated with instructions on how to export
an example into MCU+ SDK

Signed-off by: Prashant Shivhare <p-shivhare@ti.com>