AM62L: Added DDR Memtester Example
This tests the DDR memory space with 17 patterns in MCU+SDK as done in
the Linux Upstream code.
Add the ddr_memtester.c for the ddr_memtester API that can be called
in 2 ways:
1. Give '-1' as argument to test the memory space with all patterns.
2. Give any other index to test memory space with the respective pattern.
Changing the 'DDR_2_MB_SIZE' macro will change the size that the pattern
is tested for. Ensure that the space used for testing does not either
cross the DDR memory region or overwrite any registers.